Update process `top` field
This field was only being updated in Smalltalk and when sleeping/awakening of the process.
The `Object>>error` code indirectly relies on this value through `ProcessFrame=>successor`.
The wrong value was making `ProcessFrame>>thisop` to return `nil` and fail the whole error handling logic. Since "message not found" errors are automatically handled by `Object>>error` (this automation can be found in `findSelectorInClass()`) this was creating an infinite loop and subsequently a stack overflow.